+++ package.json
@@ -6,7 +6,6 @@
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/react-router-dom": "^5.1.3",
-    "@types/react-tabs": "^2.3.1",
     "ajv": "^6.12.1",
     "flat": "^5.0.0",
     "lodash": "^4.17.15",
@@ -63,6 +62,7 @@
   "homepage": ".",
   "devDependencies": {
     "@types/lodash": "^4.14.150",
+    "@types/react-tabs": "^2.3.1",
     "gh-pages": "^2.2.0"
   }
 }
