+++ src/assets/schemas/config.schema.json
@@ -12,7 +12,13 @@
     "pages": {
       "type": "array",
       "items": {
-        "$ref": "#/definitions/page"
+        "$ref": "#/definitions/resource"
+      }
+    },
+    "resources": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/resource"
       }
     },
     "baseUrl": {
@@ -41,8 +47,7 @@
     }
   },
   "required": [
-    "name",
-    "pages"
+    "name"
   ],
   "definitions": {
     "errorMessageDataPath": {
@@ -104,13 +109,13 @@
         }
       }
     },
-    "page": {
+    "resource": {
       "type": "object",
       "additionalProperties": false,
       "properties": {
         "name": {
           "type": "string",
-          "description": "The name of the page. This will be presented in the menu."
+          "description": "The name of the resource. This will be presented in the menu."
         },
         "id": {
           "type": "string",
@@ -126,10 +131,10 @@
         "methods": {
           "$ref": "#/definitions/methods"
         },
-        "subResourcePages": {
+        "subResources": {
           "type": "array",
           "items": {
-            "$ref": "#/definitions/subResourcePage"
+            "$ref": "#/definitions/subResource"
           }
         },
         "customActions": {
@@ -148,15 +153,13 @@
         "id"
       ]
     },
-    "subResourcePage": {
+    "subResource": {
       "allOf": [
         {
-          "$ref": "#/definitions/page"
+          "$ref": "#/definitions/resource"
         },
         {
-          "properties": {
-            "parentResource": {}
-          }
+          "properties": {}
         }
       ]
     },
