+++ src/components/app.tsx
@@ -2,13 +2,14 @@ import React, { useEffect, useState } from 'react';
 import { HashRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import { ToastContainer, toast } from 'react-toastify';
 import ConfigService from '../services/config.service';
-import { IConfig, IConfigPage } from '../common/models/config.model';
+import { IConfig, IConfigResource } from '../common/models/config.model';
 import { Page } from '../components/page/page.comp';
 import { DetailPage } from './detailPage/detailPage.comp';
 import { Navigation } from '../components/navigation/navigation.comp';
 import { AppContext } from './app.context';
 import HttpService from '../services/http.service';
 import { CustomStyles } from './customStyles/customStyles.comp';
+import { routesHelpers } from '../helpers/routes.helpers';
 
 import './app.scss';
 import 'react-toastify/dist/ReactToastify.css';
@@ -23,30 +24,38 @@ function changeFavicon(src: string) {
   link.rel = 'shortcut icon';
   link.href = src;
   if (oldLink) {
-   document.head.removeChild(oldLink);
+    document.head.removeChild(oldLink);
   }
   document.head.appendChild(link);
- }
+}
 
 function App() {
   const [firstLoad, setFirstLoad] = useState<boolean>(true);
   const [config, setConfig] = useState<IConfig | null>(null);
-  const [activePage, setActivePage] = useState<IConfigPage | null>(config?.pages?.[0] || null);
-  const [activePathVars, setActivePathVars] = useState<{[key: string]: string}>({});
+  const [activePage, setActivePage] = useState<IConfigResource | null>(config?.resources?.[0] || config?.pages?.[0] || null);
+  const [activePathVars, setActivePathVars] = useState<{ [key: string]: string }>({});
   const [error, setError] = useState<string | null>(null);
   const [activeItem, setActiveItem] = useState<any | null>(null);
 
   const appName: string = config?.name || defaultAppName;
-  
+
   async function loadConfig(url?: string): Promise<void> {
     try {
       const windowConfig = (window as any).RESTool?.config;
-      let remoteConfig: IConfig; 
+      let remoteConfig: IConfig;
       // Try to load config from window object first
       if (!url && windowConfig) {
         remoteConfig = Object.assign({}, windowConfig, {});
       } else {
         remoteConfig = url ? await ConfigService.getRemoteConfig(url) : await ConfigService.loadDefaultConfig();
+      }
+
+      if (remoteConfig.pages && remoteConfig.resources) {
+        throw new Error('Configuration is invalid, use either resources or pages (deprecated), not both');
+      }
+
+      if (remoteConfig.pages) {
+        remoteConfig.resources = remoteConfig.pages;
       }
 
       // Setting global config for httpService
@@ -63,7 +72,6 @@ function App() {
       if (remoteConfig?.remoteUrl) {
         return await loadConfig(remoteConfig.remoteUrl);
       }
-
       setConfig(remoteConfig);
     } catch (e) {
       console.error('Could not load config file', e);
@@ -74,10 +82,10 @@ function App() {
 
   function scrollToTop(scrollDuration: number = 250) {
     var cosParameter = window.scrollY / 2,
-    scrollCount = 0,
-    oldTimestamp = performance.now();
+      scrollCount = 0,
+      oldTimestamp = performance.now();
 
-    function step (newTimestamp: number) {
+    function step(newTimestamp: number) {
       scrollCount += Math.PI / (scrollDuration / (newTimestamp - oldTimestamp));
 
       if (scrollCount >= Math.PI) {
@@ -111,51 +119,47 @@ function App() {
     }
   }, [config]);
 
-  const mainRoutes: string[] = config?.pages?.map(p => `/${p.id}`) || [];
-  const detailRoutes: string[] = config?.pages?.reduce((acc: string[], p) => {
-    const detailPageId = p.methods?.getSingle?.detailPage?.id;
-    if(detailPageId){
-      acc.push(`/${detailPageId}`);
-    } 
-    return acc;
-  }, []) || [];
-  
+  const detailPagesConfig = routesHelpers.detailRoutesConfig(config?.resources);
+
+  const mainRoutes: string[] = config?.resources?.map(p => `/${p.id}`) || [];
+  const detailRoutes: string[] = detailPagesConfig.map(conf => conf.route);
+
   return (
     <div className="restool-app">
       {
         !config ?
-        <div className="app-error">
-          {firstLoad ? 'Loading Configuration...' : 'Could not find config file.'}
-        </div> :
-        <AppContext.Provider value={{ config, activePage, setActivePage, activePathVars, setActivePathVars, activeItem, setActiveItem, error, setError, httpService }}>
-          {
-            config.customStyles &&
-            <CustomStyles
-              styles={config.customStyles}
-            />
-          }
-          <Router>
-            <aside>
-              <h1 title={appName} onClick={() => scrollToTop()}>{appName}</h1>
-              {
-                <Navigation />
-              }
-            </aside>
+          <div className="app-error">
+            {firstLoad ? 'Loading Configuration...' : 'Could not find config file.'}
+          </div> :
+          <AppContext.Provider value={{ config, detailPagesConfig, activePage, setActivePage, activePathVars, setActivePathVars, activeItem, setActiveItem, error, setError, httpService }}>
             {
-              config &&
-              <Switch>
-                <Route exact path={mainRoutes} component={Page} />
-                <Route exact path={detailRoutes} component={DetailPage} />
-                <Redirect path="/" to={`/${config?.pages?.[0]?.id || '1'}`} />
-              </Switch>
+              config.customStyles &&
+              <CustomStyles
+                styles={config.customStyles}
+              />
             }
-            <ToastContainer
-              position={toast.POSITION.TOP_CENTER}
-              autoClose={4000}
-              draggable={false}
-            />
-          </Router>
-        </AppContext.Provider>
+            <Router>
+              <aside>
+                <h1 title={appName} onClick={() => scrollToTop()}>{appName}</h1>
+                {
+                  <Navigation />
+                }
+              </aside>
+              {
+                config &&
+                <Switch>
+                  <Route exact path={detailRoutes} component={DetailPage} />
+                  <Route exact path={mainRoutes} component={Page} />
+                  <Redirect path="/" to={`/${config?.resources?.[0]?.id || '1'}`} />
+                </Switch>
+              }
+              <ToastContainer
+                position={toast.POSITION.TOP_CENTER}
+                autoClose={4000}
+                draggable={false}
+              />
+            </Router>
+          </AppContext.Provider>
       }
     </div>
   );
